+++ src/client/cypress/e2e/editor/layerform.cy.js
@@ -27,7 +27,7 @@ describe("Tests for the layer form.", () => {
           .last()
           .scrollIntoView()
           .click();
-        cy.wait("@stratigraphy_layer_edit_patch");
+        cy.wait("@update-layer");
       });
 
     cy.get('[aria-multiselectable="true"]')
@@ -39,7 +39,7 @@ describe("Tests for the layer form.", () => {
           .find('[role="option"]')
           .eq(1)
           .click();
-        cy.wait("@stratigraphy_layer_edit_patch");
+        cy.wait("@update-layer");
       });
 
     const multipleDropdownValues = [];