+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Updated layer management to use the .NET API.
+
 ## v2.0.506 - 2023-12-21
 
 ### Added